requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
spacy==3.7.4
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.5.0
plotly==5.20.0
openpyxl==3.1.2
joblib==1.4.0
